Move files to follow folder structuring
