+++ packages/synapse-interface/components/Portfolio/components/SingleNetworkPortfolio.tsx
@@ -8,11 +8,11 @@ import {
   sortTokensByBalanceDescending,
 } from '@/utils/actions/fetchPortfolioBalances'
 import { Chain } from '@/utils/types'
-import { PortfolioAccordion } from './components/PortfolioAccordion'
-import { PortfolioConnectButton } from './components/PortfolioConnectButton'
+import { PortfolioAccordion } from './PortfolioAccordion'
+import { PortfolioConnectButton } from './PortfolioConnectButton'
 import { EmptyPortfolioContent } from './PortfolioContent'
 import { FetchState } from '@/slices/portfolio/actions'
-import { PortfolioTokenAsset } from './components/PortfolioTokenAsset'
+import { PortfolioTokenAsset } from './PortfolioTokenAsset'
 import { QuestionMarkCircleIcon } from '@heroicons/react/outline'
 import { WarningMessage } from '../../Warning'
 import { TWITTER_URL, DISCORD_URL } from '@/constants/urls'
